Agregar policy de pesos

+++ mascotas-api/app/Http/Controllers/WeightController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\Pets\WeightRequest;
+use App\Repositories\PetRepository;
 use App\Repositories\WeightRepository;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -11,9 +12,10 @@ class WeightController extends Controller
 {
     protected $weightRepository;
 
-    public function __construct(WeightRepository $weightRepository)
+    public function __construct(WeightRepository $weightRepository, PetRepository $petRepository)
     {
         $this->weightRepository = $weightRepository;
+        $this->petRepository = $petRepository;
     }
 
     /**
@@ -24,6 +26,9 @@ class WeightController extends Controller
      */
     public function getWeightsByPet($pet_id): JsonResponse
     {
+        $pet = $this->petRepository->find($pet_id);
+        $this->authorize('viewMedicalDetails', $pet);
+
         $weights = $this->weightRepository->getWeightsByPet($pet_id)->values();
         return response()->json([
             'data' => compact('weights'),
@@ -56,6 +61,7 @@ class WeightController extends Controller
     public function findWeight($weight_id): JsonResponse
     {
         $weight = $this->weightRepository->find($weight_id);
+        $this->authorize('view', $weight);
 
         return response()->json([
             'data' => compact('weight')
@@ -71,6 +77,9 @@ class WeightController extends Controller
      */
     public function updateWeight(Request $request, $weight_id): JsonResponse
     {
+        $weight = $this->weightRepository->find($weight_id);
+        $this->authorize('update', $weight);
+
         $weight = $this->weightRepository->update($weight_id, $request->get('date'), $request->get('weight'));
 
         return response()->json([
@@ -87,6 +96,9 @@ class WeightController extends Controller
      */
     public function deleteWeight($weight_id): JsonResponse
     {
+        $weight = $this->weightRepository->find($weight_id);
+        $this->authorize('delete', $weight);
+
         $this->weightRepository->delete($weight_id);
 
         return response()->json([
